compiler warning fix

+++ tests/server/sws.c
@@ -18,7 +18,7 @@
  * This software is distributed on an "AS IS" basis, WITHOUT WARRANTY OF ANY
  * KIND, either express or implied.
  *
- * $Id: sws.c,v 1.98 2007-02-14 00:28:01 yangtse Exp $
+ * $Id: sws.c,v 1.99 2007-02-14 13:46:08 yangtse Exp $
  ***************************************************************************/
 
 /* sws.c: simple (silly?) web server
@@ -460,7 +460,7 @@ void storerequest(char *reqbuf)
   if (reqbuf == NULL)
     return;
 
-  totalsize = strlen(reqbuf);
+  totalsize = (ssize_t)strlen(reqbuf);
   if (totalsize == 0)
     return;
 
@@ -475,8 +475,8 @@ void storerequest(char *reqbuf)
 
   writeleft = totalsize;
   do {
-    written = fwrite((void *) &reqbuf[totalsize-writeleft],
-                     1, (size_t)writeleft, dump);
+    written = (ssize_t)fwrite((void *) &reqbuf[totalsize-writeleft],
+                              1, (size_t)writeleft, dump);
     if (written > 0)
       writeleft -= written;
   } while ((writeleft > 0) && ((error = errno) == EINTR));
